Applied localStorage -> storage refactor
