+++ __tests__/dependencies.spec.ts
@@ -21,7 +21,7 @@
 
 import getAuthFetcherWithDependencies from "../src/dependencies";
 import AuthFetcher from "../src/AuthFetcher";
-import { StorageMock } from "../src/localStorage/__mocks__/Storage";
+import { StorageMock } from "../src/storage/__mocks__/Storage";
 
 describe("dependencies", () => {
   it("performs dependency injection", () => {
