+++ __tests__/login/oidc/TokenRequester.spec.ts
@@ -25,7 +25,7 @@ import {
   IssuerConfigFetcherMock,
   IssuerConfigFetcherFetchConfigResponse
 } from "../../../src/login/oidc/__mocks__/IssuerConfigFetcher";
-import { StorageUtilityMock } from "../../../src/localStorage/__mocks__/StorageUtility";
+import { StorageUtilityMock } from "../../../src/storage/__mocks__/StorageUtility";
 import {
   DpopHeaderCreatorMock,
   DpopHeaderCreatorResponse
