+++ __tests__/login/oidc/redirectHandler/AuthCodeRedirectHandler.spec.ts
@@ -20,7 +20,7 @@
  */
 
 import "reflect-metadata";
-import { StorageUtilityMock } from "../../../../src/localStorage/__mocks__/StorageUtility";
+import { StorageUtilityMock } from "../../../../src/storage/__mocks__/StorageUtility";
 import AuthCodeRedirectHandler from "../../../../src/login/oidc/redirectHandler/AuthCodeRedirectHandler";
 import { RedirectorMock } from "../../../../src/login/oidc/__mocks__/Redirector";
 import { SessionCreatorMock } from "../../../../src/solidSession/__mocks__/SessionCreator";
