+++ __tests__/login/oidc/redirectHandler/TokenSaver.spec.ts
@@ -23,7 +23,7 @@ import "reflect-metadata";
 import TokenSaver from "../../../../src/login/oidc/redirectHandler/TokenSaver";
 import { SessionCreatorMock } from "../../../../src/solidSession/__mocks__/SessionCreator";
 import { JoseUtilityMock } from "../../../../src/jose/__mocks__/JoseUtility";
-import { StorageUtilityMock } from "../../../../src/localStorage/__mocks__/StorageUtility";
+import { StorageUtilityMock } from "../../../../src/storage/__mocks__/StorageUtility";
 
 /**
  * Test for TokenSaver
